Split BrightSpaceGradeSyncService along its seams (#1116)
The ~1055-line file is now four focused files:

- BrightSpaceGradeSelection.swift — the pure grade-selection logic:
  StudentGrade, bestGradeForStudent, suiteTotalPoints, scaledGradePush,
  roundedGradePoints. Doc comments kept verbatim; suiteTotalPoints and
  roundedGradePoints stay private to the file.

- BrightSpaceGradeClears.swift — queued grade removals: GradeClearTarget,
  processPendingGradeClears, runGradeClear, studentHasGradeSource, now
  methods on GradeSyncSweep so the db/application/classlist-cache
  threading disappears.

- BrightSpaceSyncSweep.swift — sweep orchestration, the per-group push,
  batch loaders, the two per-sweep cache actors, GradePushKey /
  GradePushTarget / GradeSyncBatchContext, and the monitor registration.
  New struct GradeSyncSweep holds the invariant dependencies (db, logger,
  application, both caches) as stored properties; resolveClient stays a
  method parameter because every caller passes a non-escaping closure
  (storing it would force @escaping into the public signature). The free
  function sweepBrightSpaceGradeSync keeps its exact signature as a thin
  wrapper, so routes, the monitor, and the tests are untouched.
  loadGradeSyncBatchContext now loads submissions and delegates the
  chunked-IN assignment/course/user loading to loadSyncContextForKeys
  instead of duplicating it.

- BrightSpaceGradeSyncService.swift — what remains: the
  BrightSpaceSyncFlaggable protocol + conformances, retry classification
  (isRetryableSyncError), and the requeue / clear / failure-recording
  flag bookkeeping.

The hand-rolled BrightSpaceGradeSyncMonitor task/start/stop class is
replaced by the shared PeriodicSweepMonitor (same registration shape as
LearnRosterReadinessSweep / LearnSectionSyncService): 60 s interval,
runImmediately: false (the shared loop already sweeps first, matching the
old loop), same credentials no-op gate, same debounce resolution, same
"pushed N grade(s)" and "BrightSpace grade sync sweep failed: …" log
lines. AppServices registers it via PeriodicSweepLifecycleHandler.

Behavior-neutral refactor; pinned by BrightSpaceGradeSyncTests,
BrightSpacePanelTests, BrightSpaceTransportTests,
BrightSpaceVerbDispatchTests, LearnRosterReconcilerTests, and the
GradeOverride suites (76 tests, all green).
